Migrate query builder View components to TypeScript
